Revert "Merge branch '34110' of github.com:mkrawczuk/node into 34110"
This reverts commit 60ab351a502760ea109b00ce7028dbd56afbb287, reversing
changes made to 47eecd27cc7f612700d7f1a623fbf426d702187d.

+++ deps/v8/src/torque/instructions.h
@@ -166,11 +166,11 @@ inline std::ostream& operator<<(std::ostream& stream,
                     << loc.GetParameterIndex() << ")";
     case DefinitionLocation::Kind::kPhi:
       return stream << "DefinitionLocation::Phi(" << std::hex
-                    << loc.GetPhiBlock() << std::dec << ", "
+                    << (uint64_t)loc.GetPhiBlock() << std::dec << ", "
                     << loc.GetPhiIndex() << ")";
     case DefinitionLocation::Kind::kInstruction:
       return stream << "DefinitionLocation::Instruction(" << std::hex
-                    << loc.GetInstruction() << std::dec << ", "
+                    << (uint64_t)loc.GetInstruction() << std::dec << ", "
                     << loc.GetInstructionIndex() << ")";
   }
 }
